Use pin tip detection fixture for tests
